Merge public into development-restricted

+++ programs/x509/load_roots.c
@@ -150,7 +150,9 @@ int main(int argc, char *argv[])
         }
         mbedtls_printf(".");
     }
-    ms = mbedtls_timing_get_timer(&timer, 0);
+    /* On 64-bit Windows and 32-bit platforms, this wraps after about
+     * 49.7 days. This shouldn't be a problem in practice. */
+    ms = (unsigned long) mbedtls_timing_get_timer(&timer, 0);
     mbedtls_printf("\n%u iterations -> %lu ms\n", opt.iterations, ms);
     exit_code = MBEDTLS_EXIT_SUCCESS;
